+++ packages/client/src/features/activeSelection.ts
@@ -4,12 +4,14 @@ import type { LanguageClient } from 'vscode-languageclient/node';
 
 export async function activate(context: vscode.ExtensionContext, languageClient: LanguageClient) {
 	await languageClient.onReady();
-	context.subscriptions.push(languageClient.onRequest(shared.GetDocumentSelectionRequest.type, () => {
+	context.subscriptions.push(languageClient.onRequest(shared.GetEditorSelectionRequest.type, () => {
 		const editor = vscode.window.activeTextEditor;
 		if (editor) {
 			return {
-				uri: languageClient.code2ProtocolConverter.asUri(editor.document.uri),
-				offset: editor.document.offsetAt(editor.selection.end),
+				textDocument: {
+					uri: languageClient.code2ProtocolConverter.asUri(editor.document.uri),
+				},
+				position: editor.selection.end,
 			};
 		}
 	}));