+++ packages/client/src/features/documentPrintWidth.ts
@@ -1,0 +1,11 @@
+import * as vscode from 'vscode';
+import * as shared from '@volar/shared';
+import type { LanguageClient } from 'vscode-languageclient/node';
+
+export async function activate(context: vscode.ExtensionContext, languageClient: LanguageClient) {
+	await languageClient.onReady();
+	context.subscriptions.push(languageClient.onRequest(shared.GetDocumentPrintWidthRequest.type, handler => {
+		const configs = vscode.workspace.getConfiguration('volar');
+		return configs.get<number>('formatting.printWidth');
+	}));
+}