+++ packages/server/src/features/customFeatures.ts
@@ -50,7 +50,6 @@ export function register(
 	connection.onNotification(shared.RestartServerNotification.type, async () => {
 		servicesManager.restartAll();
 	});
-	connection.onRequest(shared.PingRequest.type, () => 'pong' as const);
 	connection.onRequest(shared.GetRefCompleteEditsRequest.type, handler => {
 		const document = documents.get(handler.textDocument.uri);
 		if (!document) return;